Odd. Removing transaction handling in setEndpointType resulted in a lot of unit test failures, so make it optional.
This needs to be revisited. For now, workaround is to have a transaction = true/false argument, which solves the
use case of both transactional and non-transactional endpoint type resolution, but we need to get back to this
and figure out why removing transactions results in a bunch of test failures.