setup dirtyflag reporting
the DIRTY flag from the backend is being periodically reported to the front-end via dbEnum.wsCategory.dirtyFlag
upon receiving the websocket message by the frontend, the "isdirty" attribute is set to reflect the status.
as "isdirty" attribute is being observed by Studio via renderer_api, a callback is triggered in order to bubble up the info to Studio.

BUG: ZAPP-445