Fix size computations for attributes.
There are two changes here:

1) For string attributes, we want to add the length to the size
unconditionally, not just in the "typeSize > 2" case.  For example,
Matter has a "location" attribute that is a string of length 2.
Before this change, we were allocating 2 byts in the attribute store
for this attribute, which is not enough to actually store a
2-character string, and indeed writing a 2-char string to the
attribute failed.

2) For external attributes, we do not want to allocate storage in the
default attribute store (which ends up unused, because we just never
look at that slot) or in the writable attribute store (which also ends
up unused because we don't count external attribute sizes when
computing our offset into that byte array).  Fixing this saves
kilobytes of .rodata and .bss for Matter.