Allocate the mimimum amount of space necessary for string default values.
emAfLoadAttributeDefaults uses emAfReadOrWriteAttribute to write the
value from defaults into the attr store, and that ends up calling
typeSensitiveMemCopy, which for strings already checks the length and
only copies data up to that length.  That means that for strings we
only need to allocate enough space in the default store for the actual
string length+value, not for all possible values that fit in the
typeSize.