Add a responseRef to commands for various queries.
Some queries in query-command included it, some did not.  This made it
pretty confusing, as a consumer, to tell when you could expect it to
be there and when you had to deal with it not being there.  This
change adds it to various queries so it's more consistently present.