Add a storage policy for attributes.
Some SDK implementations want to force some specific storage policy
for attributes depending on implementation decisions.  Add some
mechanisms for doing that:

* Add a way to force all lists to have an "attributeAccessInterface"
  storage policy.
* Add a way to provide a map of attributes that have an
  "attributeAccessInterface" storage policy.

For now this policy is mapped to "External" storage for endpoint
attributes, but later on I want to add another value to that enum and
use it instead of "External", because there are some optimizations we
can do for "attributeAccessInterface" that we can't do for things that
are just "External".