Fix attributeAccessInterfaceAttributes validity checking.
The setup where we collected up the known attributes while parsing XML
files does not work right, because during XML file loading we skip
parsing XML files that are already loaded into the database, if there
is an existing database.

Instead, just compare attributeAccessInterfaceAttributes to the
contents of the database.