Modify helper endpoint config 4byte default (#501)
Allow for size of default value size to be configurable

defaultValue in the EmberAfDefaultOrMinMaxAttributeValue is uint16_t. For targets that have 32-bit addresses, there is wasted space in the union. We can let defaultValue be uint32_t and save on rodata region, by fitting the default value in the meta data structure itself. This allows use to specify the size of default value, if size option is not provided we default to legacy 2 bytes.