More fixes for command masks in endpoint config.
We were claiming a command is "incoming" or "outgoing" even if it was
not enabled, if it was mandatory.

This led consumers that looked at that state to be out of sync with
other consumers that just looked at the actual enabled state.

This PR makes sure we treat the actual enabled state, not theoretical
mandatory-or-not status, as the source of truth.