Fix default values for long strings.
Long strings store their length in 2 bytes at the front of the string, with the
bytes being a little-endian representation of the string length.

But the endpoint config output was spitting out the length as big-endian, not
little-endian.  This caused things to be pretty confused and led to buffer
overruns trying to read data that was not there.