Swap out the ugprade action that does nothing with a genuine --noop.
This is accomodating for a bug in a certain driving software that
wants to execute zap as a part of the upgrade, yet there is no way
to configure it in a way where you tell it "there is nothing to upgrade"
without even calling zap.

So it calls zap, and then errors with timeouts. The quick no-op
is here to essentially make that software back off quickly and move
on.