studio/uc integration feature: disable uc component automatically
When running embedded as part of Studio, ZAP will enable the
corresponding UC component if an ZCL cluster server/client is enabled.
In reverse, similar behavior will apply.

If there are only a single endpoint, disabling a ZCL cluster will cause
an UC component to be disabled.

If there are multiple endpoints, disabling the last enabled cluster
across all endpoints will trigger the corresponding UC component to be
disabled.