Fix provisional handling for Darwin to be consistent. (#1186)
We treated things with neither "introduced" nor "provisional" annotations as
provisional in availabilityHelper(), but isProvisional() returned false for
them.  This could lead us to generate non-provisional code (conditioned on
isProvisional()) that would try to call into things with provisional
availability, which would fail to compile.

The fix is to make isProvisional the source of truth for whether things are
provisional, move the "no introduced annotation" check into isProvisional, and
factor out the "have provisional annotation" logic so we can keep having the
warning when we take the "provisional because not otherwise annotated" codepath.